+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
   * Added the `api::projects::releases::ProjectReleases` endpoint to list all
     releases for a project.
+  * Added tags related api endpoints under `api::projects::repository::tags`
   * Listing commits in a repository can now be done via `Commits`
   * Added asynchronous API for query `api::AsyncQuery` and client `api::AsyncClient`.
   * Added asynchronous client `AsyncGitlab` (created by `GitlabBuilder::build_async`).